Renamed the class XmlObjectPasert to DynamicXmlParser, so that it's different from the Namespace.
