+++ trunk/src/XmlToObjectParser.Tests/Properties/AssemblyInfo.cs
