+++ trunk/src/XmlToObjectParser.Tests/XmlToObjectParser.Tests.csproj
