+++ trunk/src/XmlToObjectParser/Properties/AssemblyInfo.cs
